+++ test/cmocka/src/audio/volume/CMakeLists.txt
@@ -13,6 +13,7 @@ add_compile_options(-DUNIT_TEST)
 
 add_library(audio_for_volume STATIC
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/volume/volume.c
+	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/volume/volume_ipc3.c
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/volume/volume_generic.c
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/volume/volume_hifi3.c
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/volume/volume_hifi4.c
